Aduit Scenario
Step-By-Step Instructions

Follow the instructions to complete each step of the activity. Then, answer the 5 questions at the end of the activity before going to the next course item to compare your work to the completed exemplar.

The following supporting materials will help you complete this activity. Keep materials open as you proceed to the next steps. 

To use the supporting materials for this course item, click the links. 

Links to supporting materials: 

Botium Toys: Scope, goals, and risk assessment report

Control categories

OR

If you don’t have a Google account, you can download the supporting materials directly from the following attachments.


Conduct the next step of the security audit by completing the controls and compliance checklist. 

To complete the checklist, open the supporting materials provided in Step 1. Then:

Review the scope, goals, and risk assessment report details, with a focus on:

The assets currently managed by the IT department

The bullet points under “Additional comments” in the Risk assessment section

Consider information provided in the scenario, the scope, goals, and risk assessment report, as well as details provided in other documents linked within the checklist. 

Then, review the question in the controls and compliance sections of the checklist and select “yes” or “no” to answer the question in each section (note: the recommendations section is optional).*

To use the supporting materials for this step, click the following link.

Link to supporting materials: 
Controls and compliance checklist
 

OR

If you don’t have a Google account, you can download the supporting materials directly from the following attachment.

*If using the DOCX File, type an X to select “yes” or “no”.



Pro Tip: Save a copy of your work
Finally, be sure to download and save a copy of your completed activity to your own device. You can upload it to the portfolio platform of your choice, then share with potential employers to help demonstrate your knowledge and experience.

What to Include in Your Response

Be sure to address the following elements in your completed activity:

Controls and compliance checklist

“Yes” or “no” is selected to answer the question related to each control listed

“Yes” or “no” is selected toanswer the question related to each compliance best practice

A recommendation is provided for the IT manager (optional)

Step 3: Assess your activity
The following is a self-assessment for your controls and compliance checklist. You will use these statements to review your own work. The self-assessment process is an important part of the learning experience because it allows you to objectively assess your security audit.

There are a total of 5 points possible for this activity and each statement is worth 1 point. The items correspond to each step you completed for the activity. 

To complete the self-assessment, first open your controls assessment and compliance checklist. Then respond yes or no to each statement. 

When you complete and submit your responses, you will receive a percentage score. This score will help you confirm whether you completed the required steps of the activity. The recommended passing grade for this project is at least 80% (or 4/5 points). If you want to increase your score, you can revise your project and then resubmit your responses to reflect any changes you made. Try to achieve at least 4 points before continuing on to the next course item.